Implement nested inputs
The implementation is based on the first example in #3. The second example could
have been implemented pretty easily but makes the code too complicated for what
it does, and it would have introduced a second way to do a single thing, which
is bad.

There should be one-- and preferably only one --obvious way to do it.

This commit closes #3.